Implement a better fix for escaped strings in referenced dictionaries
When merging reference dictionaries with escaped strings, these strings
would be double processed. The original fix simply did not wrap strings
in Value objects when creating the parameter objects during the merge
step for dictionaries in the ValueList.render method.

This change replaces that fix, which did not work in all cases, by
not parsing string during the creation of parameter objects in
ValueList.render, instead strings are simply wraped as ScalarItems.
This is valid as any strings will have already been wraped once and
rendered so can not be references.